Dog spawns from pocketWatch (kinda, should spawn the way the sheep did)
Will do that soon

Edited the VoxelMat TEXTURE for the Faded mat, so that it has a darker side so the voxels from the prop stand out a little more.
Still need the player's prop to be their frame of reference

Made the collider on the pocket watch smaller so less random stuff may hit it.